requirements.txt
# === Inteligencia Artificial (Gemini) ===
google-generativeai>=0.8.3

# === Google Calendar API ===
google-api-python-client>=2.149.0
google-auth-oauthlib>=1.2.1
google-auth>=2.35.0

# === Calendarios (ICS / iCalendar) ===
ics>=0.7
icalendar>=5.0.12

# === Utilidades ===
python-dotenv>=1.0.1
PyYAML>=6.0.2
tomli; python_version < '3.11'  # compatibilidad si usas Python 3.10
requests>=2.32.3

# === Tipos y helpers ===
types-PyYAML>=6.0.12.12

protobuf~=6.33.0